* Scanina maksima

+++ shop/src/main/java/com/tofi/shop/controller/CartController.java
@@ -7,6 +7,7 @@ import com.tofi.shop.service.ItemService;
 import com.tofi.shop.service.ServiceException;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistration;
 
 import javax.inject.Inject;
 import java.util.List;
@@ -35,7 +36,7 @@ public class CartController {
 
     @PostMapping("/add")
     public @ResponseBody String addItemToCart(@RequestParam(value = "id", defaultValue = "-1") String itemId ) throws ServiceException {
-        long id = Long.parseLong(itemId);
+        Integer id = Integer.parseInt(itemId);
         if (id == -1) return "FAIL";
         Item item = this.itemService.findById(id);
         this.cartService.addItem(item);
