+++ shop/src/main/java/com/tofi/shop/dao/ItemDAO.java
@@ -6,6 +6,6 @@ import com.tofi.shop.domain.Item;
  * Data Access Object interface with methods
  * for accessing {@link Item}s in the database.
  */
-public interface ItemDAO extends GenericDAO<Long,Item> {
+public interface ItemDAO extends GenericDAO<Integer,Item> {
 
 }
